requirements.txt
aiogram==3.4.1
apscheduler==3.10.4
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9
